Relocate methods
